Polish LSP example
